requirements.txt
fastapi
uvicorn
bleak
jinja2
python-multipart
httpx
winrt-runtime==3.2.1; sys_platform == 'win32'
